# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    API specification. Requests to an unknown route under `/api` return a 404
    response using the `Error` schema with the message `API route not found`.
    Requests with malformed JSON bodies return a 400 response using the `Error`
    schema with the message `Malformed JSON request body`. Unexpected server
    failures return a generic 500 response using the same schema. JSON request
    bodies that exceed the parser's size limit return a 413 response using the
    `Error` schema with the message `Request body too large`.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions:
    get:
      operationId: listPublishedQuestions
      tags: [questions]
      summary: List published questions
      description: Returns only published questions matching the learner's session filters.
      parameters:
        - $ref: "#/components/parameters/Domain"
        - $ref: "#/components/parameters/Approach"
        - $ref: "#/components/parameters/Difficulty"
      responses:
        "200":
          description: Published questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /instructor/questions:
    get:
      operationId: listInstructorQuestions
      tags: [instructor]
      summary: List all questions for the instructor workspace
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/QuestionStatus"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Domain"
        - $ref: "#/components/parameters/Approach"
        - $ref: "#/components/parameters/Difficulty"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: Instructor question library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorQuestionPage"
        "400":
          description: Invalid instructor question filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createQuestion
      tags: [instructor]
      summary: Create a question draft
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "200":
          description: Existing question returned for an idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Idempotency key was reused with different question content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: JSON request body exceeds the size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /instructor/questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [instructor]
      summary: Update a question draft or published question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: JSON request body exceeds the size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /instructor/questions/{id}/publish:
    post:
      operationId: publishQuestion
      tags: [instructor]
      summary: Publish a question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Published question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /instructor/questions/{id}/archive:
    post:
      operationId: archiveQuestion
      tags: [instructor]
      summary: Archive a question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Archived question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /instructor/questions/{id}/restore:
    post:
      operationId: restoreQuestion
      tags: [instructor]
      summary: Restore an archived question as a draft
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Restored draft question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Question is not archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique UUID for this create operation; reuse it when retrying after an ambiguous failure.
      schema:
        type: string
        format: uuid
...
    QuestionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    Domain:
      name: domain
      in: query
      schema:
        $ref: "#/components/schemas/QuestionDomain"
    Approach:
      name: approach
      in: query
      schema:
        $ref: "#/components/schemas/QuestionApproach"
    Difficulty:
      name: difficulty
      in: query
      schema:
        $ref: "#/components/schemas/QuestionDifficulty"
    Search:
      name: search
      in: query
      description: Case-insensitive substring matched against the question topic or text.
      schema:
        type: string
        minLength: 1
    PageLimit:
      name: limit
      in: query
      description: Maximum number of questions to return. Defaults to 25 and cannot exceed 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    PageOffset:
      name: offset
      in: query
      description: Number of matching questions to skip before the current page.
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    QuestionDomain:
      type: string
      enum: [People, Process, "Business environment"]
    QuestionApproach:
      type: string
      enum: [agile, predictive, hybrid]
    QuestionDifficulty:
      type: string
      enum: [easy, medium, hard]
    QuestionStatus:
      type: string
      enum: [draft, published, archived]
    Question:
      type: object
      required:
        - id
        - domain
        - topic
        - approach
        - difficulty
        - question
        - translation
        - options
        - correctAnswer
        - explanation
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        domain:
          $ref: "#/components/schemas/QuestionDomain"
        topic:
          type: string
          minLength: 1
        approach:
          $ref: "#/components/schemas/QuestionApproach"
        difficulty:
          $ref: "#/components/schemas/QuestionDifficulty"
        question:
          type: string
          minLength: 1
        translation:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: string
            minLength: 1
        correctAnswer:
          type: integer
          minimum: 0
        explanation:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/QuestionStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InstructorQuestionPagination:
      type: object
      required: [offset, limit, total, hasNext, nextOffset]
      properties:
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        nextOffset:
          type: ["integer", "null"]
          minimum: 0
    InstructorQuestionPage:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        pagination:
          $ref: "#/components/schemas/InstructorQuestionPagination"
    QuestionInput:
      type: object
      required:
        - domain
        - topic
        - approach
        - difficulty
        - question
        - translation
        - options
        - correctAnswer
        - explanation
      properties:
        domain:
          $ref: "#/components/schemas/QuestionDomain"
        topic:
          type: string
          minLength: 1
        approach:
          $ref: "#/components/schemas/QuestionApproach"
        difficulty:
          $ref: "#/components/schemas/QuestionDifficulty"
        question:
          type: string
          minLength: 1
        translation:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 6
          items:
            type: string
            minLength: 1
        correctAnswer:
          type: integer
          minimum: 0
        explanation:
          type: string
          minLength: 1
    QuestionUpdate:
      allOf:
        - $ref: "#/components/schemas/QuestionInput"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string